OPNClient: accept arbitrary request parameters
This allows for prototyping new clients
that use request parameters that are not defined in OPNClient (yet).

This should be regarded as an advanced feature that may break on misuse.
For instance, passing 'data=...' to _post() will raise
a TypeError since the 'body' parameter gets translated to that already.
Cannot change that without breaking existing client classes.